+++ DEMO/Form/Form1.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace DEMO
+{
+    public partial class Form_DangNhap : Form
+    {
+        public Form_DangNhap()
+        {
+            InitializeComponent();
+        }
+
+
+    }
+}
