+++ DEMO/Program.cs
@@ -16,7 +16,7 @@ namespace DEMO
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new Form_QuanLy());
         }
     }
 }
